fix(xs-oracle): stop truncating the differential regexp match meter (fuzz 5d122a6fc10babd9)
The continuous-Ironhorse-fuzz `differential_regexp` arm flagged a "match
meter" divergence on a 3-byte input (sha256 79942bf9…d7a917) that folds
into a deeply nested empty-matchable pattern over "00b00"@2. Matching it
dispatches 243671 metered steps, so the raw 16.16 match meter is
243671*65536 = 15_969_222_656 — past 2^32.

Both engines dispatch the identical step histogram (verified op-for-op),
so the port's meter was never wrong. XS's own `meterIndex` is a txU8
(xsAll.h); the port meters into a `u64` and holds the full value. The
oracle shim, however, copied that meter into a 32-bit `txU4` field and so
wrapped it to 15_969_222_656 mod 2^32 = 3_084_320_768 — a phantom
divergence against the port's correct figure. Same shape as the prior
standing finding (493390fc03979205): a harness-side truncation, not an
engine bug.

Fix (oracle-side only): widen the regexp meter fields — compile/match
raw meters and computrons — to 64 bit end to end (EndorRegExpResult in
xs_shim.c and XsRegExpResultRaw/RegExpOutcome in lib.rs), and drop the now
-redundant `as u64` casts at the two comparison sites.

Regressions:
- xs-oracle unit test `regexp_match_meter_is_not_truncated_to_32_bits`:
  asserts the oracle reports the full 15_969_222_656 (fails pre-fix with
  the wrapped value). Needs the submodule, like the sibling oracle tests.
- ironhorse-vm `finding_5d122a6fc10babd9_regexp_meter_overflow`: replays
  the exact case through the port's `compile`+`match_regexp`, asserting no
  panic and the full-width meter (> u32::MAX). Builds without the oracle
  or the c/moddable submodule, per the standing-branch convention.

+++ rust/engine/ironhorse-fuzz/src/regexp.rs
@@ -263,7 +263,7 @@ pub fn differential_check_regexp(case: &RegExpCase) -> Result<bool, Divergence> 
             });
         }
     }
-    if outcome.match_meter_raw != oracle.match_meter_raw as u64 {
+    if outcome.match_meter_raw != oracle.match_meter_raw {
         return Err(Divergence {
             source,
             detail: format!(
